+++ tool_packages/unique_internal_search/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.17] - 2026-01-13
+- Removing unused parameters from config
+- Activating multiple search string execution by default
+
 ## [1.2.16] - 2026-01-12
 - Fix message logswith new ChatUI