+++ unique_orchestrator/CHANGELOG.md
@@ -5,8 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.7] - 2026-01-13
+- Removing unused and retired configuration parameters `max_review_steps`, `uploaded_content_config`, `tool_progress_reporter_config`, `force_checks_on_stream_response_references` and `thinking_steps_display`
+
 ## [1.11.6] - 2026-01-13
-- Deprecate UniqueAIResponsesApi and use UniqueAI directly instead (type safety) 
+- Deprecate UniqueAIResponsesApi and use UniqueAI directly instead (type safety)
 
 ## [1.11.5] - 2026-01-12
 - Include feature flag to have message logs compatible with new ChatUI