+++ unique_toolkit/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.43.3] - 2026-01-13
+- Removing unused configuration `UploadedContentConfig` from `history manager`
+- Changing default `assessment_type` in `SubAgentEvaluationServiceConfig` to `HALLUCINATION`
+
 ## [1.43.2] - 2026-01-12
 - `DocxGeneratorService`: Alignment need to be specified by the template rather than in code